maint: move modal and delete project logic
